Issue number: Obtained from: Submitted by:  mranga Reviewed by:   mranga
Remove 1.5 specific constructs from code ( should all run under 1.4 )
CVS: ----------------------------------------------------------------------
CVS: Issue number:
CVS:   If this change addresses one or more issues,
CVS:   then enter the issue number(s) here.
CVS: Obtained from:
CVS:   If this change has been taken from another system,
CVS:   then name the system in this line, otherwise delete it.
CVS: Submitted by:
CVS:   If this code has been contributed to the project by someone else; i.e.,
CVS:   they sent us a patch or a set of diffs, then include their name/email
CVS:   address here. If this is your work then delete this line.
CVS: Reviewed by:
CVS:   If we are doing pre-commit code reviews and someone else has
CVS:   reviewed your changes, include their name(s) here.
CVS:   If you have not had it reviewed then delete this line.

+++ src/gov/nist/javax/sip/SipStackImpl.java
@@ -191,7 +191,7 @@ import gov.nist.core.net.NetworkLayer;
  * 
  * 
  * 
- * @version 1.2 $Revision: 1.49 $ $Date: 2006-11-02 04:06:14 $
+ * @version 1.2 $Revision: 1.50 $ $Date: 2006-11-02 15:26:31 $
  * 
  * @author M. Ranganathan <br/>
  * 
@@ -543,7 +543,7 @@ public class SipStackImpl extends SIPTransactionStack implements
 		if (interval != null) {
 			try {
 				// Make the monitored threads ping the auditor twice as fast as the audits
-				getThreadAuditor().setPingIntervalInMillisecs(Long.valueOf(interval) / 2);
+				getThreadAuditor().setPingIntervalInMillisecs(Long.valueOf(interval).longValue() / 2);
 			} catch (NumberFormatException ex) {
 				System.out.println("THREAD_AUDIT_INTERVAL_IN_MILLISECS - bad value [" + interval + "] " + ex.getMessage());
 			}
